+++ app/src/main/java/com/example/sasha/MapPageFragment.java
@@ -25,15 +25,8 @@ import com.naver.maps.map.OnMapReadyCallback;
 import com.naver.maps.map.overlay.Marker;
 import com.naver.maps.map.util.FusedLocationSource;
 
-public class MapPageFragment extends Fragment implements OnMapReadyCallback {
-    private static final String TAG = "map";
-    private static final int PERMISSION_REQUEST_CODE = 1000;
-    private static final String[] PERMISSIONS = {
-            Manifest.permission.ACCESS_FINE_LOCATION,
-            Manifest.permission.ACCESS_COARSE_LOCATION
-    };
-    private FusedLocationSource mLocationSource;
-    private NaverMap mNaverMap;
+public class MapPageFragment extends Fragment {
+
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
@@ -41,54 +34,11 @@ public class MapPageFragment extends Fragment implements OnMapReadyCallback {
         MainActivity mainActivity = (MainActivity) getActivity();
         ViewGroup rootView = (ViewGroup) inflater.inflate(R.layout.fragment_map_page, container, false);
 
-        FragmentManager fm = getChildFragmentManager();
-        MapFragment mapFragment = (MapFragment) fm.findFragmentById(R.id.map);
-        if (mapFragment == null) {
-            mapFragment = MapFragment.newInstance();
-            fm.beginTransaction().add(R.id.map, mapFragment).commit();
-        }
-        mapFragment.getMapAsync(MapPageFragment.this);
-        mLocationSource = new FusedLocationSource(this, PERMISSION_REQUEST_CODE);
 
-
-
-
-        Button btn_tmp = rootView.findViewById(R.id.btn_tmp);
-        btn_tmp.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(mainActivity.getApplicationContext(), LoginActivity.class);
-                startActivity(intent);
-            }
-        });
 
           return rootView;
     }
 
 
-    @Override
-    public void onMapReady(@NonNull NaverMap naverMap) {
-        Log.d(TAG, "OnMapReady");
 
-//        Marker marker = new Marker();
-//        marker.setPosition(new LatLng(37.5670135, 126.9783740));
-//        marker.setMap(naverMap);
-
-        mNaverMap = naverMap;
-        mNaverMap.setLocationSource(mLocationSource);
-
-        MainActivity mainActivity = (MainActivity) getActivity();
-        ActivityCompat.requestPermissions(mainActivity, PERMISSIONS, PERMISSION_REQUEST_CODE);
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        if (requestCode == PERMISSION_REQUEST_CODE) {
-            if (grantResults.length > 0 &&
-                    grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                mNaverMap.setLocationTrackingMode(LocationTrackingMode.Follow);
-            }
-        }
-    }
 }
